fix(billing): extract SettingsLayout to component — Next.js disallows named page exports
