+++ frontend/app/settings/page.tsx
@@ -2,53 +2,7 @@
 
 import { useEffect } from "react";
 import { useRouter } from "next/navigation";
-import Link from "next/link";
-import AppShell from "@/components/AppShell";
 import { useAuth } from "@/components/auth/AuthProvider";
-import { isAdmin, isOwner } from "@/lib/auth";
-import { pageTitle } from "@/lib/styles";
-
-const tabs = [
-  { href: "/settings", label: "Profile", minRole: "member" as const },
-  { href: "/settings/security", label: "Security", minRole: "member" as const },
-  { href: "/settings/organization", label: "Organization", minRole: "admin" as const },
-  { href: "/settings/billing", label: "Billing", minRole: "owner" as const },
-];
-
-function SettingsLayout({ children, activeTab }: { children: React.ReactNode; activeTab: string }) {
-  const { user } = useAuth();
-  if (!user) return null;
-
-  const visibleTabs = tabs.filter((tab) => {
-    if (tab.minRole === "owner") return isOwner(user);
-    if (tab.minRole === "admin") return isAdmin(user);
-    return true;
-  });
-
-  return (
-    <AppShell>
-      <h1 className={pageTitle}>Settings</h1>
-      <div className="mb-6 flex gap-0 border-b border-border">
-        {visibleTabs.map((tab) => (
-          <Link
-            key={tab.href}
-            href={tab.href}
-            className={`px-5 py-3 text-sm font-medium transition-colors ${
-              activeTab === tab.href
-                ? "border-b-2 border-accent text-accent"
-                : "text-text-muted hover:text-text-primary"
-            }`}
-          >
-            {tab.label}
-          </Link>
-        ))}
-      </div>
-      {children}
-    </AppShell>
-  );
-}
-
-export { SettingsLayout };
 
 export default function SettingsProfilePage() {
   const { user } = useAuth();
